py-backend/pyproject.toml
[project]
name = "attendance-backend"
version = "0.1.0"
description = "Smart attendance backend with Flask, SCRFD and face embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easydict>=1.13",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "insightface>=0.7.3",
    "mmcv==2.1.0",
    "mmdet>=3.3.0",
    "ninja>=1.13.0",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "pytorch-lightning>=2.6.1",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "setuptools>=81.0.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "wheel>=0.46.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
